Added unit tests for DepartmentServiceImpl, updated DepartmentServiceImpl functionality

+++ src/main/java/com/wisetechglobal/employees/service/DepartmentServiceImpl.java
@@ -5,11 +5,15 @@ import com.wisetechglobal.employees.persistence.repository.DepartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Service
 public class DepartmentServiceImpl implements DepartmentService {
+
+    private final String MSG_TPL_ENTITY_DOESNT_EXIST = "Department with id=%d doesn't exist";
 
     private final DepartmentRepository departmentRepository;
 
@@ -31,7 +35,11 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Override
     public Department updateDepartment(final Department department, final Integer id) {
 
-        Department departmentDb = departmentRepository.findById(id).get();
+        Optional<Department> result = departmentRepository.findById(id);
+        if (result.isEmpty()) {
+            throw new EntityNotFoundException(String.format(MSG_TPL_ENTITY_DOESNT_EXIST, id));
+        }
+        Department departmentDb = result.get();
         if (Objects.nonNull(department.getName())) {
             departmentDb.setName(department.getName());
         }
@@ -40,6 +48,10 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public void deleteDepartmentById(Integer id) {
+        Optional<Department> result = departmentRepository.findById(id);
+        if (result.isEmpty()) {
+            throw new EntityNotFoundException(String.format(MSG_TPL_ENTITY_DOESNT_EXIST, id));
+        }
         departmentRepository.deleteById(id);
     }
 }
